Factor out scenario
